requirements.txt
Django==1.11.18
djangorestframework==3.9.1
